libraryValidations/Python/requirements.txt
pytest
featuremanagement["AzureMonitor"]==2.0.0b3
azure-appconfiguration-provider==2.0.0b3
